scripts/requirements.txt
beautifulsoup4==4.5.1
Requests==2.11.1
pandas==0.18.1
python-magic==0.4.12
pdfminer3k==1.3.1
